requirements.txt
aniso8601==9.0.1
black==22.1.0
click==8.0.3
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-decouple==3.6
python-http-client==3.3.5
pytz==2021.3
sendgrid==6.9.5
six==1.16.0
SQLAlchemy==1.4.31
starkbank-ecdsa==2.0.3
tomli==2.0.0
typing-extensions==4.0.1
Werkzeug==2.0.2
